refactor(metrics): finalize public API for publishing
