+++ Cargo.toml
@@ -44,8 +44,8 @@ anyhow = "1.0.102"
 backtrace = "0.3.76"
 foundations = { version = "5.8.1", path = "./foundations", default-features = false }
 foundations-macros = { version = "=5.8.1", path = "./foundations-macros", default-features = false }
-foundations-metrics-registry = { version = "0.1.0", path = "./foundations-metrics-registry", default-features = false }
-foundations-metrics = { version = "0.1.0", path = "./foundations-metrics", default-features = false }
+foundations-metrics-registry = { version = "1.0.0", path = "./foundations-metrics-registry", default-features = false }
+foundations-metrics = { version = "0.1.0-beta1", path = "./foundations-metrics", default-features = false }
 bindgen = { version = "0.72.1", default-features = false }
 cc = "1.2.61"
 cf-rustracing = "1.4.0"
