+++ foundations-metrics-registry/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "foundations-metrics-registry"
 description = "Process-global metric registry and Prometheus protobuf data model for Foundations."
-version = "0.1.0"
+version = "1.0.0"
 edition = { workspace = true }
 repository = { workspace = true }
 authors = { workspace = true }
